cleanup the cli and other files that got big and bloated
