+++ service/src/cli/mod.rs
@@ -1,0 +1,9 @@
+pub mod framework_tool;
+
+// Back-compat re-exports for existing imports: crate::cli::{FrameworkTool, resolve_or_install}
+pub use framework_tool::{
+    FrameworkTool,
+    resolve_or_install,
+};
+
+
