+++ service/src/update.rs
@@ -1,5 +1,5 @@
-use serde_json::Value;
 use tracing::{error, info};
+use crate::utils::github as gh;
 
 pub fn parse_github_repo_env() -> Option<(String, String)> {
     let repo = std::env::var("FRAMEWORK_CONTROL_UPDATE_REPO").ok()?;
@@ -21,51 +21,13 @@ pub fn parse_github_repo_env() -> Option<(String, String)> {
     }
 }
 
-pub async fn fetch_latest_release(owner: &str, name: &str) -> Result<Value, String> {
-    let api = format!("https://api.github.com/repos/{}/{}/releases/latest", owner, name);
-    let resp = reqwest::Client::new()
-        .get(api)
-        .header("user-agent", "framework-control-service")
-        .send()
-        .await
-        .map_err(|e| e.to_string())?;
-    let text = resp.text().await.map_err(|e| e.to_string())?;
-    serde_json::from_str::<Value>(&text).map_err(|e| e.to_string())
-}
-
-pub fn extract_latest_version_tag(parsed: &Value) -> Option<String> {
-    let tag = parsed.get("tag_name").and_then(|v| v.as_str())?;
-    let v = tag.trim_start_matches('v').to_string();
-    if v.is_empty() { None } else { Some(v) }
-}
-
-pub fn find_installer_url(parsed: &Value) -> Option<String> {
-    let assets = parsed.get("assets")?.as_array()?.clone();
-    #[cfg(target_os = "windows")]
-    let preferred_exts: &[&str] = &[".msi"];
-    #[cfg(target_os = "macos")]
-    let preferred_exts: &[&str] = &[".pkg", ".dmg"];
-    #[cfg(all(unix, not(target_os = "macos")))]
-    let preferred_exts: &[&str] = &[".deb", ".rpm", ".AppImage"];
-
-    assets.iter().find_map(|a| {
-        let name = a.get("name").and_then(|v| v.as_str()).unwrap_or("");
-        let matches = preferred_exts.iter().any(|ext| name.ends_with(ext));
-        if matches {
-            a.get("browser_download_url").and_then(|v| v.as_str()).map(|s| s.to_string())
-        } else {
-            None
-        }
-    })
-}
 
 pub async fn get_current_and_latest() -> Result<(String, Option<String>), String> {
     let current = env!("CARGO_PKG_VERSION").to_string();
     let Some((owner, name)) = parse_github_repo_env() else {
         return Ok((current, None));
     };
-    let parsed = fetch_latest_release(&owner, &name).await?;
-    let latest = extract_latest_version_tag(&parsed);
+    let latest = gh::get_latest_release_version_tag(&owner, &name).await?;
     Ok((current, latest))
 }
 
@@ -99,19 +61,21 @@ pub async fn check_and_apply_now() -> Result<bool, String> {
     let Some((owner, name)) = parse_github_repo_env() else {
         return Err("FRAMEWORK_CONTROL_UPDATE_REPO not set".into());
     };
-    let current = env!("CARGO_PKG_VERSION").to_string();
-    let parsed = fetch_latest_release(&owner, &name).await.map_err(|e| {
-        error!("update: fetch latest failed: {}", e);
-        e
-    })?;
-    let latest = match extract_latest_version_tag(&parsed) {
-        Some(v) => v,
-        None => return Ok(false),
-    };
+    let (current, latest_opt) = get_current_and_latest().await?;
+    let latest = match latest_opt { Some(v) => v, None => return Ok(false) };
     if latest <= current {
         return Ok(false);
     }
-    let Some(installer_url) = find_installer_url(&parsed) else {
+    #[cfg(target_os = "windows")]
+    let preferred_exts: &[&str] = &[".msi"];
+    #[cfg(target_os = "macos")]
+    let preferred_exts: &[&str] = &[".pkg", ".dmg"];
+    #[cfg(all(unix, not(target_os = "macos")))]
+    let preferred_exts: &[&str] = &[".deb", ".rpm", ".AppImage"];
+    let Some(installer_url) = gh::get_latest_release_url_ending_with(&owner, &name, preferred_exts).await.map_err(|e| {
+        error!("update: fetch assets failed: {}", e);
+        e
+    })? else {
         error!("update: no installer asset in latest release");
         return Err("installer asset not found".into());
     };
