+++ service/src/utils/mod.rs
@@ -1,0 +1,5 @@
+pub mod github;
+pub mod wget;
+pub mod download;
+
+
